refactor : 공통 로직 분리하여 모듈화
