+++ pages/sign/sign.js
@@ -5,8 +5,6 @@ export const emailInput = document.querySelector('#sign-email');
 export const passwordInput = document.querySelector('#sign-password');
 export const passwordEyeButton = document.querySelector('.eye-off');
 
-const submitButton = document.querySelector('.btn.login');
-
 export const emailErrorMessageElement = document.createElement("p");
 export const passwordErrorMessageElement = document.createElement("p");
 
@@ -74,24 +72,8 @@ export function toggleEyeButton(input, eyeButton) {
     }
 }
 
-function login(e) {
-    e.preventDefault();
-
-    const email = emailInput.value;
-    const password = passwordInput.value;
-
-    if (email === "test@codeit.com" && password === "codeit101") {
-        location.href = "/folder.html";
-        return;
-    }
-
-    showErrorMessage(emailInput, emailErrorMessageElement, errorMessages.email.invalid);
-    showErrorMessage(passwordInput, passwordErrorMessageElement, errorMessages.password.invalid);
-}
-
 emailInput.addEventListener("focusout", validateEmailType);
 passwordInput.addEventListener("focusout", validatePassword);
-submitButton.addEventListener("click", login);
 passwordEyeButton.addEventListener("click", () => {
     toggleEyeButton(passwordInput, passwordEyeButton);
 });
