+++ pages/sign/signin.html
@@ -56,6 +56,6 @@
       </div>
     </div>
   </main>
-<script type="module" src="sign.js"></script>
+<script type="module" src="sign-in.js"></script>
 </body>
 </html>
